+++ src/main/kotlin/com/rankweis/uppercut/karate/psi/formatter/KarateJavascriptFormat.java
@@ -3,7 +3,6 @@ package com.rankweis.uppercut.karate.psi.formatter;
 import com.intellij.application.options.CodeStyle;
 import com.intellij.formatting.Alignment;
 import com.intellij.formatting.Block;
-import com.intellij.formatting.Formatter;
 import com.intellij.formatting.FormattingMode;
 import com.intellij.formatting.Indent;
 import com.intellij.formatting.Wrap;
@@ -12,8 +11,11 @@ import com.intellij.lang.ASTNode;
 import com.intellij.lang.javascript.JSLanguageDialect;
 import com.intellij.lang.javascript.formatter.JSBlockContext;
 import com.intellij.lang.javascript.formatter.JSCodeStyleSettings;
-import com.intellij.psi.TokenType;
+import com.intellij.lang.javascript.formatter.blocks.JSBlock;
+import com.intellij.lang.javascript.formatter.blocks.SubBlockVisitor;
+import com.intellij.psi.PsiFile;
 import com.intellij.psi.codeStyle.CodeStyleSettings;
+import com.intellij.psi.codeStyle.CommonCodeStyleSettings;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -29,55 +31,49 @@ public class KarateJavascriptFormat {
   public List<Block> getJsSubBlocks(ASTNode astNode, Alignment alignment) {
     List<Block> result = new ArrayList<>();
     final ASTNode[] children = astNode.getChildren(null);
-    CodeStyleSettings settings = CodeStyle.getSettings(astNode.getPsi().getContainingFile());
-    JSBlockContext context = new JSBlockContext(settings, dialect, null,
-      FormattingMode.REFORMAT);
-      Wrap wrap;
-    Wrap childWrap = null;
-    context.getCommonSettings().SPACE_BEFORE_METHOD_PARENTHESES = false;
-    context.getDialectSettings().SPACE_BEFORE_FUNCTION_LEFT_PARENTH = false;
-    context.getCommonSettings().SPACE_BEFORE_METHOD_PARENTHESES = false;
-    if (!astNode.getText().contains("\n")) {
-      wrap = Wrap.createWrap(WrapType.NONE, false);
-      childWrap = Formatter.getInstance().createChildWrap(wrap, WrapType.NONE, false);
-      JSCodeStyleSettings commonSettings = JSCodeStyleSettings.getSettings(astNode.getPsi());
-      commonSettings.VAR_DECLARATION_WRAP = 0;
-      commonSettings.OBJECT_LITERAL_WRAP = 0;
-      commonSettings.OBJECT_TYPES_WRAP = 0;
-      commonSettings.IMPORTS_WRAP = 0;
-      commonSettings.UNION_TYPES_WRAP = 0;
-      commonSettings.BLANK_LINES_AROUND_FUNCTION = 0;
-      CodeStyleSettings defaults = CodeStyleSettings.getDefaults();
-      defaults.WRAP_LONG_LINES = false;
-      defaults.METHOD_ANNOTATION_WRAP = 0;
-      defaults.CLASS_ANNOTATION_WRAP = 0;
-      defaults.FIELD_ANNOTATION_WRAP = 0;
-      defaults.KEEP_SIMPLE_METHODS_IN_ONE_LINE = true;
-      defaults.BLANK_LINES_AROUND_METHOD = 0;
+    PsiFile file = astNode.getPsi().getContainingFile();
+    CodeStyleSettings settings = CodeStyle.getSettings(file);
+    CodeStyle.runWithLocalSettings(file.getProject(), settings, (s) -> {
 
-      context =
-        new JSBlockContext(defaults, dialect, commonSettings,
-          FormattingMode.REFORMAT);
-      context.getDialectSettings().IMPORTS_WRAP = 0;
-      context.getDialectSettings().OBJECT_LITERAL_WRAP = 0;
-      context.getDialectSettings().VAR_DECLARATION_WRAP = 0;
-      context.getDialectSettings().BLANK_LINES_AROUND_FUNCTION = 0;
-      context.getCommonSettings().FIELD_ANNOTATION_WRAP = 0;
-      context.getCommonSettings().METHOD_ANNOTATION_WRAP = 0;
-      context.getCommonSettings().WRAP_LONG_LINES = false;
-      context.getCommonSettings().METHOD_ANNOTATION_WRAP = 0;
-      context.getCommonSettings().CLASS_ANNOTATION_WRAP = 0;
-      context.getCommonSettings().FIELD_ANNOTATION_WRAP = 0;
-      context.getCommonSettings().KEEP_SIMPLE_METHODS_IN_ONE_LINE = true;
-    }
+      JSCodeStyleSettings explicitSettings =
+        s.getCustomSettings(JSCodeStyleSettings.getSettingsClass(dialect.getBaseLanguage())).clone();
+      CommonCodeStyleSettings commonCodeStyleSettings = s.getCommonSettings(dialect.getBaseLanguage());
 
-    for (ASTNode jsChild : children) {
-      if (jsChild.getElementType() == TokenType.WHITE_SPACE) {
-        continue;
+      commonCodeStyleSettings.SPACE_WITHIN_BRACES = true;
+      explicitSettings.SPACE_BEFORE_FUNCTION_LEFT_PARENTH = false;
+      commonCodeStyleSettings.SPACE_BEFORE_METHOD_PARENTHESES = false;
+
+      Wrap childWrap = null;
+      if (!astNode.getText().contains("\n")) {
+        commonCodeStyleSettings.WRAP_LONG_LINES = false;
+        commonCodeStyleSettings.KEEP_SIMPLE_METHODS_IN_ONE_LINE = true;
+        commonCodeStyleSettings.KEEP_SIMPLE_BLOCKS_IN_ONE_LINE = true;
+        commonCodeStyleSettings.KEEP_SIMPLE_LAMBDAS_IN_ONE_LINE = true;
+        commonCodeStyleSettings.KEEP_SIMPLE_CLASSES_IN_ONE_LINE = true;
+        commonCodeStyleSettings.KEEP_MULTIPLE_EXPRESSIONS_IN_ONE_LINE = true;
+
+        explicitSettings.SPACES_WITHIN_OBJECT_LITERAL_BRACES = true;
+        explicitSettings.SPACES_WITHIN_OBJECT_TYPE_BRACES = true;
+        commonCodeStyleSettings.SPACE_WITHIN_BRACES = true;
+      } else {
+        childWrap = Wrap.createWrap(WrapType.NORMAL, false);
       }
-      Block b = context.createBlock(jsChild, childWrap, alignment, Indent.getNormalIndent(), null, null);
-      result.add(b);
-    }
+
+      JSBlockContext context =
+        new KarateJavaScriptBlockContext(s, dialect.getBaseLanguage(), explicitSettings, FormattingMode.REFORMAT);
+      JSBlock parent = new KarateJavascriptBlock(astNode, alignment, Indent.getNoneIndent(), childWrap,
+        (x) -> x.getTreeParent() == astNode ? alignment : null, context);
+      SubBlockVisitor subBlockVisitor = context.createSubBlockVisitor(parent, null);
+      result.add(parent);
+//      for (ASTNode jsChild : children) {
+//        if (jsChild.getElementType() == TokenType.WHITE_SPACE) {
+//          subBlockVisitor.visit(jsChild);
+//        }
+//        Block b = context.createBlock(jsChild, childWrap, alignmentChild, Indent.getNoneIndent(), null, parent);
+//        result.add(b);
+//      }
+//      result.addAll(subBlockVisitor.getBlocks());
+    });
     return result;
   }