+++ src/main/kotlin/com/rankweis/uppercut/karate/psi/impl/GherkinPsiElementBase.java
@@ -14,14 +14,12 @@ import com.intellij.psi.util.PsiUtilCore;
 import com.rankweis.uppercut.karate.psi.GherkinElementVisitor;
 import com.rankweis.uppercut.karate.psi.GherkinPsiElement;
 import com.rankweis.uppercut.karate.psi.KarateTokenTypes;
-import groovy.transform.EqualsAndHashCode;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import javax.swing.Icon;
 import org.jetbrains.annotations.NotNull;
 
-@EqualsAndHashCode
 public abstract class GherkinPsiElementBase extends ASTWrapperPsiElement implements GherkinPsiElement {
 
   private static final TokenSet TEXT_FILTER = TokenSet.create(KarateTokenTypes.TEXT);
@@ -93,13 +91,27 @@ public abstract class GherkinPsiElementBase extends ASTWrapperPsiElement impleme
     }
   }
 
-  protected abstract void acceptGherkin(GherkinElementVisitor gherkinElementVisitor);
-
-  @Override public boolean equals(Object o) {
-    if (!(o instanceof GherkinPsiElementBase that)) {
-      return false;
+  @Override
+  public boolean equals(Object obj) {
+    if (this == obj) {
+      return true; // Same instance
     }
-    return Objects.equals(getNode().getText(), that.getNode().getText()) && Objects.equals(getNode().getTextRange(),
-      that.getNode().getTextRange()) && getContainingFile().equals(that.getContainingFile());
+    if (!(obj instanceof GherkinPsiElementBase other)) {
+      return false; // Different types
+    }
+
+    ASTNode thisNode = this.getNode();
+    ASTNode otherNode = other.getNode();
+
+    if (thisNode == null || otherNode == null) {
+      return false; // One or both nodes are missing
+    }
+
+    return thisNode.getElementType().equals(otherNode.getElementType()) &&
+      thisNode.getPsi().getContainingFile() == other.getNode().getPsi().getContainingFile() &&
+      thisNode.getTextRange().equals(otherNode.getTextRange()) &&
+      thisNode.getText().equals(otherNode.getText());
   }
+
+  protected abstract void acceptGherkin(GherkinElementVisitor gherkinElementVisitor);
 }